REF/CLN: move around cache utilities
